Updated README and Makefile
The README is still outdated, but at least the first sections are better
now. The intro and setup is a bit better, and the Makefile got a tweak.

+++ Makefile
@@ -17,7 +17,7 @@ setup-unix:
 	@# ln -sf $(CURDIR)/environment ~/.kde/env/environment.sh
 	@# This config is for git to modify at will, so is only copied. Is where the
 	@# user name and email can be set to each ones values.
-	cp -f $(CURDIR)/gitconfig ~/.gitconfig
+	test -f ~/.gitconfig || cp -f $(CURDIR)/gitconfig ~/.gitconfig
 	@# This is not changed by "git config --global", so it can be under version
 	@# control, and improved by hand like the other files.
 	ln -sf $(CURDIR)/gitconfig.extra ~/.gitconfig.extra
